calculate tempo related numbers on higher precision. Huge thanks @georgkrause - that's a difficult bug you just fixed!

+++ src/looper.hxx
@@ -45,7 +45,7 @@ public:
 	void setRequestedBuffer(int s, AudioBuffer* ab);
 
 	/// stores the framesPerBeat from TimeManager. Used to stretch audio
-	void setFpb(int f);
+	void setFpb(double f);
 
 	/// Retrieve a clip from the Looper
 	LooperClip* getClip(int scene);
@@ -67,7 +67,7 @@ private:
 	int playingScene;
 	int queuedScene;
 
-	int fpb;
+	double fpb;
 
 	//vector<float> tmpRecordBuffer;
 	LooperClip* clips[NSCENES];
